+++ projects/ngx-view-state/src/lib/view-state/factories/error-view-status.ts
@@ -1,9 +1,0 @@
-import { ViewStatusEnum } from '../models/view-status.enum';
-import { ViewError } from '../models/view-status.model';
-
-export function errorViewStatus<E>(error?: E): ViewError<E> {
-  return {
-    type: ViewStatusEnum.ERROR,
-    error,
-  };
-}