+++ projects/ngx-view-state/src/lib/view-state/factories/idle-view-status.ts
@@ -1,8 +1,0 @@
-import { ViewStatusEnum } from '../models/view-status.enum';
-import { ViewIdle } from '../models/view-status.model';
-
-export function idleViewStatus(): ViewIdle {
-  return {
-    type: ViewStatusEnum.IDLE,
-  };
-}