+++ projects/ngx-view-state/src/lib/view-state/services/view-state-actions.service.spec.ts
@@ -13,8 +13,8 @@ describe('ViewStateActionsService', () => {
   const actionsConfig: ViewStateActionsConfig[] = [
     {
       startLoadingOn: loadData,
-      resetLoadingOn: [loadDataSuccess],
-      error: [lodDataFailure],
+      resetOn: [loadDataSuccess],
+      errorOn: [lodDataFailure],
     },
   ];