+++ projects/ngx-view-state/src/lib/view-state/store/view-state.integration.spec.ts
@@ -42,8 +42,8 @@ describe('ViewStateIntegration', () => {
       this.viewStateActionsService.add([
         {
           startLoadingOn: DataActions.loadData,
-          resetLoadingOn: [DataActions.loadDataSuccess],
-          error: [DataActions.loadDataFailure],
+          resetOn: [DataActions.loadDataSuccess],
+          errorOn: [DataActions.loadDataFailure],
         },
       ]);
     }