+++ projects/ngx-view-state/CHANGELOG.md
@@ -1,5 +1,12 @@
 ## Changelog
 
+## 2.0.0    
+
+- Rename ViewStateActionsConfig properties
+  - `resetLoadingOn` to `resetOn`
+  - `error` to `errorOn`
+
+
 ## 1.0.2
 
 - Update peerDependencies and keywords